Edit/Update individual jots functionality
This commit features additions to complete the edit/update jot
functionality. The capability requires two interactions - edit button,
which reveals a form to edit the jot fields, and save, which actually
updates the api and performs a PATCH request.

Bug: Upon PATCH, the updated row gets pushed to the bottom of the
table, thus throwing the displayed jots out of order. I've opened a
ticket on this already:
https://git.generalassemb.ly/ga-wdi-boston/full-stack-project/issues/1546